+++ packages/file/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Dart 3 fixes for class modifiers.
 * `MemoryFileSystem` now treats empty paths as non-existent.
 * Fix `FileSystem.isLink`/`FileSystem.isLinkSync` to not follow symbolic links.
+* Make the return type of `MemoryFile.openRead` and `_ChrootFile.openRead` again
+  match the return type from `dart:io`.
 
 #### 6.1.4